Moved plotting from lsa_service mc_service and sde_model_inversion_service.
