+++ examples/main_vep.py
@@ -159,9 +159,10 @@ def main_vep(test_write_read=False, pse_flag=PSE_FLAG, sa_pse_flag=SA_PSE_FLAG, 
                                                      os.path.join(FOLDER_RES, hyp.name + "_ModelConfig.h5")),
                                                  logger=logger)))
         # Plot nullclines and equilibria of model configuration
-        model_configuration_service.plot_state_space(model_configuration, head.connectivity.region_labels,
-                                                     special_idx=disease_indices, model="6d", zmode="lin",
-                                                     figure_name=hyp.name + "_StateSpace")
+        plotter = Plotter()
+        plotter.plot_state_space(model_configuration, head.connectivity.region_labels,
+                                 special_idx=disease_indices, model="6d", zmode="lin",
+                                 figure_name=hyp.name + "_StateSpace")
 
         logger.info("\n\nRunning LSA...")
         lsa_service = LSAService(eigen_vectors_number=None, weighted_eigenvector_sum=True)
@@ -177,7 +178,8 @@ def main_vep(test_write_read=False, pse_flag=PSE_FLAG, sa_pse_flag=SA_PSE_FLAG, 
                 assert_equal_objects(lsa_hypothesis,
                                      reader.read_hypothesis(os.path.join(FOLDER_RES, lsa_hypothesis.name + "_LSA.h5")),
                                      logger=logger)))
-        lsa_service.plot_lsa(lsa_hypothesis, model_configuration, head.connectivity.region_labels, None)
+        plotter.plot_lsa(lsa_hypothesis, model_configuration, lsa_service.weighted_eigenvector_sum,
+                         lsa_service.eigen_vectors_number, head.connectivity.region_labels, None)
         if pse_flag:
             # --------------Parameter Search Exploration (PSE)-------------------------------
             logger.info("\n\nRunning PSE LSA...")
@@ -190,7 +192,8 @@ def main_vep(test_write_read=False, pse_flag=PSE_FLAG, sa_pse_flag=SA_PSE_FLAG, 
                                                       {"name": "x0_values", "indices": healthy_indices}],
                                                   model_configuration_service=model_configuration_service,
                                                   lsa_service=lsa_service, logger=logger, save_flag=True)[0]
-            lsa_service.plot_lsa(lsa_hypothesis, model_configuration, head.connectivity.region_labels, pse_results)
+            plotter.plot_lsa(lsa_hypothesis, model_configuration, lsa_service.weighted_eigenvector_sum,
+                             lsa_service.eigen_vectors_number, head.connectivity.region_labels, pse_results)
             # , show_flag=True, save_flag=False
             writer.write_dictionary(pse_results, os.path.join(FOLDER_RES, lsa_hypothesis.name + "_PSE_LSA_results.h5"))
             if test_write_read:
@@ -213,8 +216,9 @@ def main_vep(test_write_read=False, pse_flag=PSE_FLAG, sa_pse_flag=SA_PSE_FLAG, 
                                                                  {"name": "x0_values", "indices": healthy_indices}],
                                                              model_configuration_service=model_configuration_service,
                                                              lsa_service=lsa_service, logger=logger)
-            lsa_service.plot_lsa(lsa_hypothesis, model_configuration, head.connectivity.region_labels, pse_sa_results,
-                                 title="SA PSE Hypothesis Overview")
+            plotter.plot_lsa(lsa_hypothesis, model_configuration, lsa_service.weighted_eigenvector_sum,
+                             lsa_service.eigen_vectors_number, head.connectivity.region_labels, pse_sa_results,
+                             title="SA PSE Hypothesis Overview")
             # , show_flag=True, save_flag=False
             writer.write_dictionary(pse_sa_results,
                                     os.path.join(FOLDER_RES, lsa_hypothesis.name + "_SA_PSE_LSA_results.h5"))
@@ -265,9 +269,9 @@ def main_vep(test_write_read=False, pse_flag=PSE_FLAG, sa_pse_flag=SA_PSE_FLAG, 
                                                                  hpf_flag=True, hpf_low=10.0, hpf_high=512.0,
                                                                  sensors_list=head.sensorsSEEG)
                 # Plot results
-                Plotter().plot_sim_results(sim.model, lsa_hypothesis.lsa_propagation_indices, vois_ts_dict,
-                                 head.sensorsSEEG, hpf_flag=True, trajectories_plot=trajectories_plot,
-                                 spectral_raster_plot=spectral_raster_plot, log_scale=True)
+                plotter.plot_sim_results(sim.model, lsa_hypothesis.lsa_propagation_indices, vois_ts_dict,
+                                         head.sensorsSEEG, hpf_flag=True, trajectories_plot=trajectories_plot,
+                                         spectral_raster_plot=spectral_raster_plot, log_scale=True)
                 # Optionally save results in mat files
                 # from scipy.io import savemat
                 # savemat(os.path.join(FOLDER_RES, lsa_hypothesis.name + "_ts.mat"), vois_ts_dict)
